Add in tonne of scripts to run jobs on HTCondor.
There is one per executable. Idea is that the user just edits

submit_XXX_condor.sh

which then submits a job (using submit_template.condor). On the
worker node, it executes XXX_condor.sh, which calls condor_wrapper.py
to deal with hdfs output (since we can't stream to hdfs). This is a
bit hacky, and should be done in one script...