Lower the default fit_min in setup_fit() in runCalibration.py.
This now takes into account 0 PU samples where there is no 'turnover'. Still
keep in the max(fit_min, max_corr_pt) though, to account for the turnover, if there is one.
If there isn't, then max_corr_pt will be the lowest pt point on the graph anyway, which will be about 10.